requirements.txt
Django>=6.0.1
djangorestframework>=3.16.1
celery>=5.6.2
redis>=5.0.0
django-celery-results>=2.6.0
pymysql>=1.1.2
django-unfold>=0.76.0
django-import-export>=4.4.0
django-guardian>=3.2.0
django-simple-history>=3.11.0
psutil>=7.2.1
Pillow>=11.1.0
# Optional / Future
# python-libtorrent
# deepl
